requirements.txt
fastapi==0.123.0
joblib==1.4.2
matplotlib==3.10.7
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.5
PyYAML==6.0.3
Requests==2.32.5
scikit_learn==1.7.2
streamlit==1.43.2
mlflow==3.6.0